LayoutTests/platform/ios-sim-deprecated/fast/forms/input-number-large-padding-expected.txt

Test for a bug that spin buttons did not work if a number input has large padding-right value.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL numberInput.value should be 1. Was 0.
PASS successfullyParsed is true

TEST COMPLETE
